BE: 분 단위 주식 데이터 조회

+++ backend/src/main/java/com/ssafy/newstock/stock/repository/StockInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface StockInfoRepository extends JpaRepository<StockInfo, Integer> {
+public interface StockInfoRepository extends JpaRepository<StockInfo, Long> {
 
     @Query("select s from StockInfo s where s.date >= :startDate and s.stockCode = :stockCode and s.period = :period")
     List<StockInfo> findDataWithinRange(@Param("startDate") String startDate,
